fix(osm2pgsql): fixed pr issues
