+++ images/osm2pgsql/requirements.txt
@@ -1,3 +1,3 @@
-psycopg2
+psycopg2==2.8.6
 mapcoloniesjsonlogger==0.0.14
-requests
+requests==2.24.0
